sejong_catch_backend/sejong_catch/sejong-auth-crawler-svc/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
PyJWT>=2.9
python-dotenv>=1.0
sejong-univ-auth>=0.3.4

# EmailStr 사용을 위해 반드시 필요
pydantic[email]>=2.7
# (중복 설치 문제 없음) 필요시 명시적으로:
email-validator>=2.1.0.post1

# TLS 핸드셰이크 이슈 회피용(맥 + Python 조합)
requests==2.31.0
urllib3==1.26.18
certifi
